Use `AutoRegisterMixin` class for `RendererBase`
